# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ExamForge — exam generator API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /exams/{examId}/questions:
    post:
      operationId: createQuestion
      tags: [questions]
      summary: Add a question to an exam
      parameters:
        - in: path
          name: examId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateQuestionBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Question" }
  /exams/{examId}/questions/bulk:
    post:
      operationId: bulkImportQuestions
      tags: [questions]
      summary: Bulk import questions into an exam from a JSON array
      parameters:
        - in: path
          name: examId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BulkImportQuestionsBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BulkImportResult" }
  /questions/{questionId}:
    patch:
      operationId: updateQuestion
      tags: [questions]
      summary: Update a question
      parameters:
        - in: path
          name: questionId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateQuestionBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Question" }
    delete:
      operationId: deleteQuestion
      tags: [questions]
      summary: Delete a question
      parameters:
        - in: path
          name: questionId
          required: true
          schema: { type: string }
      responses:
        "204": { description: Deleted }
...
  /exams/{examId}/leaderboard:
    get:
      operationId: getLeaderboard
      tags: [leaderboard]
      summary: Get leaderboard for an exam
      parameters:
        - in: path
          name: examId
          required: true
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/LeaderboardEntry" }
    post:
      operationId: submitLeaderboardEntry
      tags: [leaderboard]
      summary: Submit a leaderboard entry
      parameters:
        - in: path
          name: examId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubmitLeaderboardEntryBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LeaderboardEntry" }
...
components:
  schemas:
...
    Question:
      type: object
      properties:
        id: { type: string }
        examId: { type: string }
        questionType: { type: string, enum: [mcq, essay], default: mcq }
        topic: { type: string, nullable: true }
        prompt: { type: string }
        options:
          type: array
          items: { type: string }
        correctIndex: { type: integer }
        essayAnswer: { type: string, nullable: true }
        explanation: { type: string, nullable: true }
        reference: { type: string, nullable: true }
        repeatNote: { type: string, nullable: true }
        position: { type: integer }
        createdAt: { type: string, format: date-time }
      required: [id, examId, questionType, prompt, options, correctIndex, position, createdAt]
    CreateQuestionBody:
      type: object
      properties:
        questionType: { type: string, enum: [mcq, essay], default: mcq }
        topic: { type: string, nullable: true }
        prompt: { type: string, minLength: 1 }
        options:
          type: array
          items: { type: string, minLength: 1 }
          minItems: 2
          maxItems: 8
        correctIndex: { type: integer, minimum: 0 }
        essayAnswer: { type: string, nullable: true }
        explanation: { type: string, nullable: true }
        reference: { type: string, nullable: true }
        repeatNote: { type: string, nullable: true }
      required: [prompt]
    BulkImportQuestionsBody:
      type: object
      properties:
        questions:
          type: array
          minItems: 1
          maxItems: 500
          items: { $ref: "#/components/schemas/CreateQuestionBody" }
      required: [questions]
    BulkImportResult:
      type: object
      properties:
        insertedCount: { type: integer }
        examId: { type: string }
      required: [insertedCount, examId]
    UpdateQuestionBody:
      type: object
      properties:
        questionType: { type: string, enum: [mcq, essay] }
        topic: { type: string, nullable: true }
        prompt: { type: string, minLength: 1 }
        options:
          type: array
          items: { type: string, minLength: 1 }
          minItems: 2
          maxItems: 8
        correctIndex: { type: integer, minimum: 0 }
        essayAnswer: { type: string, nullable: true }
        explanation: { type: string, nullable: true }
        reference: { type: string, nullable: true }
        repeatNote: { type: string, nullable: true }
        position: { type: integer }
...
    LeaderboardEntry:
      type: object
      properties:
        id: { type: string }
        examId: { type: string }
        userName: { type: string }
        score: { type: integer }
        total: { type: integer }
        scorePct: { type: number }
        elapsedSeconds: { type: integer }
        completedAt: { type: string, format: date-time }
      required: [id, examId, userName, score, total, scorePct, elapsedSeconds, completedAt]
    SubmitLeaderboardEntryBody:
      type: object
      properties:
        userName: { type: string, minLength: 1 }
        score: { type: integer, minimum: 0 }
        total: { type: integer, minimum: 1 }
        elapsedSeconds: { type: integer, minimum: 0 }
      required: [userName, score, total]
...