# lib/api-zod/src/generated/types/bulkImportQuestionsBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ExamForge — exam generator API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateQuestionBody } from "./createQuestionBody";

export interface BulkImportQuestionsBody {
  /**
   * @minItems 1
   * @maxItems 500
   */
  questions: CreateQuestionBody[];
}