# lib/api-zod/src/generated/types/createQuestionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ExamForge — exam generator API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateQuestionBodyQuestionType } from "./createQuestionBodyQuestionType";

export interface CreateQuestionBody {
  questionType?: CreateQuestionBodyQuestionType;
  topic?: string | null;
  /** @minLength 1 */
  prompt: string;
  /**
   * @minItems 2
   * @maxItems 8
   */
  options?: string[];
  /** @minimum 0 */
  correctIndex?: number;
  essayAnswer?: string | null;
  explanation?: string | null;
  reference?: string | null;
  repeatNote?: string | null;
}